Update simCategorical.R
Hello,

I have an error, when simulation with the multion method
> Error in nnet.default(X, Y, w, mask = mask, size = 0, skip = TRUE, softmax = TRUE, :
too many (4284) weights

It is caused by low MaxNWts parameter. I do not know anything about the parameter. The only thing that interests me is that I need it higher to run the model. https://stackoverflow.com/questions/17105979/i-get-error-error-in-nnet-defaultx-y-w-too-many-77031-weights-whi So increased the value.